Allow packages to install phpunit
Instead of using a "replace" which prevents the phpunit package from
being installed no mater where in the stack this package lives, we allow
Yoast to require "phpunit/phpunit" and cleanup the extra files when the
package is installed as a standalone.

Opens greater possibilities for using this package in projects, which
include their own phpunit versions while still allowing a clean
standalone version with external phpunit by way of .phar.